Fixed some shit

+++ Assets/Scripts/dbController.cs
@@ -141,6 +141,7 @@ public class dbController : MonoBehaviour
         }
         else if (questionID != 0)
         {
+            int imgID2 = 0;
             dbconn = new SqliteConnection("URI=file:" + Application.dataPath + "/database/Database.s3db");
             dbconn.Open();
 
@@ -148,20 +149,28 @@ public class dbController : MonoBehaviour
             cmd.CommandText = "SELECT * FROM Vraag WHERE VraagID=" + questionID;
             SqliteDataReader reader = cmd.ExecuteReader();
 
-            int imgID2 = Convert.ToInt32(reader[1]);
-            reader.Close();
-            try
+            if (reader[1] != DBNull.Value)
             {
-                deleteRects(imgID2);
-            }
-            catch (Exception e)
-            {
-                Debug.Log("Exception thrown: " + e.Message);
-            }
+                Debug.Log(@"Reader[1] != null!");
+                imgID2 = Convert.ToInt32(reader[1]);
+                reader.Close();
+                try
+                {
+                    deleteRects(imgID2);
+                }
+                catch (Exception e)
+                {
+                    Debug.Log("Exception thrown: " + e.Message);
+                }
 
-            
-            cmd.CommandText = "DELETE FROM Afbeelding WHERE AfbeeldingID=" + imgID2;
-            cmd.ExecuteScalar();
+
+                cmd.CommandText = "DELETE FROM Afbeelding WHERE AfbeeldingID=" + imgID2;
+                cmd.ExecuteScalar();
+            }
+            else
+            {
+                Debug.Log("Could not find imageID associated with the questionID " + questionID);
+            }
 
             dbconn.Close();
         }
@@ -595,7 +604,7 @@ public class dbController : MonoBehaviour
         if (correct == true) 
             cmd.Parameters.Add(new SqliteParameter("@correct", 1));
         else
-            cmd.Parameters.Add(new SqliteParameter("@correct", 0));
+            cmd.Parameters.Add(new SqliteParameter("@correct", -1));
         cmd.ExecuteNonQuery();
 
         dbconn.Close();
@@ -664,12 +673,20 @@ public class dbController : MonoBehaviour
         SqliteCommand cmd = new SqliteCommand();
 
         cmd.Connection = dbconn;
-        cmd.CommandText = "SELECT * FROM Antwoord WHERE Vraag=" + question;
+        cmd.CommandText = "SELECT AntwoordID FROM Antwoord, Vraag WHERE Antwoord.VraagID = Vraag.VraagID AND Vraag='" + question + "'";
         SqliteDataReader reader = cmd.ExecuteReader();
 
         while (reader.Read())
         {
-            answers.Add((int)((Int32)reader[0]));
+            if (reader[0] == DBNull.Value)
+            {
+                answers.Add(0);
+            }
+            else
+            {
+                answers.Add(Convert.ToInt32(reader[0]));
+            }
+            
         }
 
         dbconn.Close();
@@ -677,7 +694,7 @@ public class dbController : MonoBehaviour
         return answers;
     }
 
-    public List<int> getAnswerIDs(int questionId)
+    public List<int> getAnswerIDs(int questionID)
     {
         List<int> answers = new List<int>();
 
@@ -687,12 +704,20 @@ public class dbController : MonoBehaviour
         SqliteCommand cmd = new SqliteCommand();
 
         cmd.Connection = dbconn;
-        cmd.CommandText = "SELECT * FROM Antwoord WHERE VraagID=" + questionId;
+        cmd.CommandText = "SELECT * FROM Antwoord WHERE VraagID=" + questionID;
         SqliteDataReader reader = cmd.ExecuteReader();
 
         while (reader.Read())
         {
-            answers.Add(Convert.ToInt32(reader[0]));
+            if (reader[0] == DBNull.Value)
+            {
+                answers.Add(0);
+            }
+            else
+            {
+                answers.Add(Convert.ToInt32(reader[0]));
+            }
+
         }
 
         dbconn.Close();
@@ -762,8 +787,17 @@ public class dbController : MonoBehaviour
         SqliteCommand cmd = new SqliteCommand();
 
         cmd.Connection = dbconn;
-        cmd.CommandText = "SELECT Correct FROM Antwoord WHERE Antwoord=" + answerStr;
-        answerINT = Convert.ToInt32(cmd.ExecuteScalar());
+        cmd.CommandText = "SELECT Correct FROM Antwoord WHERE Antwoord='" + answerStr + "'";
+        object result = cmd.ExecuteScalar();
+
+        if (result == DBNull.Value)
+        {
+            answerINT = 0;
+        }
+        else
+        {
+            answerINT = Convert.ToInt32(result);
+        }
 
         if (answerINT > 0)
             answer = true;
